Refactor war simulation into modular package
