+++ simulation/war/systems/__init__.py
@@ -1,0 +1,13 @@
+from systems.combat import CombatSystem
+from systems.movement import MovementSystem
+from systems.moral import MoralSystem
+from systems.pathfinding import PathfindingSystem
+from systems.victory import VictorySystem
+
+__all__ = [
+    "CombatSystem",
+    "MovementSystem",
+    "MoralSystem",
+    "PathfindingSystem",
+    "VictorySystem",
+]
