+++ simulation/war/terrain/__init__.py
@@ -1,0 +1,17 @@
+from tools.terrain_generators import (
+    carve_river,
+    generate_base,
+    place_forest,
+    place_lake,
+    place_mountains,
+    place_swamp_desert,
+)
+
+__all__ = [
+    "carve_river",
+    "generate_base",
+    "place_forest",
+    "place_lake",
+    "place_mountains",
+    "place_swamp_desert",
+]
