+++ simulation/war/ui/__init__.py
@@ -1,0 +1,3 @@
+from systems.pygame_viewer import PygameViewerSystem
+
+__all__ = ["PygameViewerSystem"]
